chore: replace redux with zustand
