+++ package.json
@@ -17,8 +17,7 @@
     "date-fns": "^4.1.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "react-redux": "^9.2.0",
-    "redux": "^5.0.1"
+    "zustand": "^5.0.4"
   },
   "devDependencies": {
     "@eslint/js": "^9.18.0",
